feat: removed dependency on jsdom for servver-side rendering
