+++ package.json
@@ -83,7 +83,6 @@
     "estrella": "^1.4.1",
     "http-server": "^14.1.1",
     "jest": "^29.0.3",
-    "jest-environment-jsdom": "^29.0.3",
     "jest-silent-reporter": "^0.5.0",
     "less": "^4.1.3",
     "open": "^8.4.0",
