+++ src/mathlive.ts
@@ -3,7 +3,6 @@ import type {
   AutoRenderOptions,
   MathfieldOptions,
   RemoteVirtualKeyboardOptions,
-  TextToSpeechOptions,
 } from './public/options';
 export * from './public/mathlive';
 
@@ -12,31 +11,20 @@ import {
   autoRenderMathInElement,
 } from './addons/auto-render';
 export * from './addons/auto-render';
-import MathLiveDebug, {
-  asciiMathToLatex,
-  latexToAsciiMath,
-} from './addons/debug';
-import { toMathML } from './addons/math-ml';
+import MathLiveDebug from './addons/debug';
 import './addons/definitions-metadata';
 
-import { atomToSpeakableText } from './editor/atom-to-speakable-text';
 import { VirtualKeyboard } from './editor/virtual-keyboard-utils';
 import './editor/virtual-keyboard-commands';
 import { RemoteVirtualKeyboard } from './editor-mathfield/remote-virtual-keyboard';
 
-import { Atom } from './core/atom-class';
-import {
-  parseLatex,
-  validateLatex as validateLatexInternal,
-} from './core/parser';
-import { adjustInterAtomSpacing, coalesce, makeStruts, Box } from './core/box';
-import { Context } from './core/context';
 import { isBrowser, throwIfNotInBrowser } from './common/capabilities';
-import { defaultGlobalContext } from './core/core';
-import { DEFAULT_FONT_SIZE } from './core/font-metrics';
-
-import { ComputeEngine, SemiBoxedExpression } from '@cortex-js/compute-engine';
-import { Expression, LatexSyntaxError } from './public/mathlive';
+import {
+  convertLatexToMarkup,
+  convertLatexToMathMl,
+  convertLatexToSpeakableText,
+  serializeMathJsonToLatex,
+} from 'public/mathlive-ssr';
 
 export type MathLiveGlobal = {
   version: string;
@@ -119,178 +107,6 @@ export function makeSharedVirtualKeyboard(
     globalMathLive().sharedVirtualKeyboard = new RemoteVirtualKeyboard(options);
   }
   return globalMathLive().sharedVirtualKeyboard!;
-}
-
-/**
- * Convert a LaTeX string to a string of HTML markup.
- *
- * **(Note)**
- *
- * This function does not interact with the DOM. The function does not load
- * fonts or inject stylesheets in the document. It can be used
- * on the server side.
- *
- * To get the output of this function to correctly display
- * in a document, use the mathlive static style sheet by adding the following
- * to the `<head>` of the document:
- *
- * ```html
- * <link rel="stylesheet" href="https://unpkg.com/mathlive/dist/mathlive-static.css" />
- * ```
- *
- * ---
- *
- * @param text A string of valid LaTeX. It does not have to start
- * with a mode token such as `$$` or `\(`.
- *
- * @param options.mathstyle If `"displaystyle"` the "display" mode of TeX
- * is used to typeset the formula, which is most appropriate for formulas that are
- * displayed in a standalone block.
- *
- * If `"textstyle"` is used, the "text" mode
- * of TeX is used, which is most appropriate when displaying math "inline"
- * with other text (on the same line).
- *
- * @param  options.macros A dictionary of LaTeX macros
- *
- *
- * @category Converting
- * @keywords convert, latex, markup
- */
-export function convertLatexToMarkup(
-  text: string,
-  options?: {
-    mathstyle?: 'displaystyle' | 'textstyle';
-    format?: string;
-  }
-): string {
-  options = options ?? {};
-  options.mathstyle = options.mathstyle ?? 'displaystyle';
-
-  const context = defaultGlobalContext();
-
-  //
-  // 1. Parse the formula and return a tree of atoms, e.g. 'genfrac'.
-  //
-  const root = new Atom('root', context);
-  root.body = parseLatex(text, context, {
-    parseMode: 'math',
-    mathstyle: options.mathstyle,
-  });
-
-  //
-  // 2. Transform the math atoms into elementary boxes
-  // for example from genfrac to VBox.
-  //
-  const box = root.render(
-    new Context(
-      {
-        registers: context.registers,
-        renderPlaceholder: () => new Box(0xa0, { maxFontSize: 1.0 }),
-      },
-      {
-        fontSize: DEFAULT_FONT_SIZE,
-        letterShapeStyle: context.letterShapeStyle,
-      },
-      options.mathstyle
-    )
-  );
-
-  if (!box) return '';
-
-  //
-  // 3. Adjust to `mord` according to TeX spacing rules
-  //
-  adjustInterAtomSpacing(box);
-
-  //
-  // 2. Simplify by coalescing adjacent boxes
-  //    for example, from <span>1</span><span>2</span>
-  //    to <span>12</span>
-  //
-  coalesce(box);
-
-  //
-  // 4. Wrap the expression with struts
-  //
-  const wrapper = makeStruts(box, { classes: 'ML__mathlive' });
-
-  //
-  // 5. Generate markup
-  //
-
-  return wrapper.toMarkup();
-}
-
-export function validateLatex(s: string): LatexSyntaxError[] {
-  return validateLatexInternal(s, defaultGlobalContext());
-}
-
-/**
- * Convert a LaTeX string to a string of MathML markup.
- *
- * @param latex A string of valid LaTeX. It does not have to start
- * with a mode token such as a `$$` or `\(`.
- * @param options.generateId If true, add an `"extid"` attribute
- * to the MathML nodes with a value matching the `atomID`. This can be used
- * to map items on the screen with their MathML representation or vice-versa.
- * @param options.onError Callback invoked when an error is encountered while
- * parsing the input string.
- *
- * @category Converting
- */
-
-export function convertLatexToMathMl(
-  latex: string,
-  options: { generateID?: boolean } = {}
-): string {
-  return toMathML(
-    parseLatex(latex, defaultGlobalContext(), {
-      parseMode: 'math',
-      args: () => '',
-      mathstyle: 'displaystyle',
-    }),
-    options
-  );
-}
-
-/**
- * Convert a LaTeX string to a textual representation ready to be spoken
- *
- * @param latex A string of valid LaTeX. It does not have to start
- * with a mode token such as a `$$` or `\(`.
- *
- * @param options {@inheritDoc TextToSpeechOptions}
- *
- * @return The spoken representation of the input LaTeX.
- * @example
- * console.log(convertLatexToSpeakableText('\\frac{1}{2}'));
- * // 'half'
- * @category Converting
- * @keywords convert, latex, speech, speakable, text, speakable text
- */
-export function convertLatexToSpeakableText(
-  latex: string,
-  options: Partial<TextToSpeechOptions> = {}
-): string {
-  const atoms = parseLatex(latex, defaultGlobalContext(), {
-    parseMode: 'math',
-    mathstyle: 'displaystyle',
-  });
-
-  return atomToSpeakableText(atoms, options as Required<TextToSpeechOptions>);
-}
-
-let gComputeEngine: ComputeEngine;
-
-export function serializeMathJsonToLatex(json: Expression): string {
-  if (!gComputeEngine) {
-    const ComputeEngineCtor =
-      globalThis[Symbol.for('io.cortexjs.compute-engine')]?.ComputeEngine;
-
-    if (ComputeEngineCtor) gComputeEngine = new ComputeEngineCtor();
-  }
-  return gComputeEngine?.box(json as SemiBoxedExpression).latex ?? '';
 }
 
 /**
@@ -379,8 +195,6 @@ export const version = {
 
 /** @internal */
 export const debug = {
-  latexToAsciiMath,
-  asciiMathToLatex,
   FUNCTIONS: MathLiveDebug.FUNCTIONS,
   MATH_SYMBOLS: MathLiveDebug.MATH_SYMBOLS,
   TEXT_SYMBOLS: MathLiveDebug.TEXT_SYMBOLS,
